refactor(ui): extract view-model/people-derivations.ts
Move deriveDuplicatePeople and deriveVisiblePeopleActors into
view-model/people-derivations.ts. These share the cleanText +
normalizeDisplayName helpers from the internal module and define the
display rules for people/actors in the People tab. The barrel
re-exports both so helpers.ts and people.ts stay on the same import
path.
